# dwengo_backend/openapi3_0.yaml
openapi: 3.0.0
info:
  title: Learning Platform API
  version: "1.0"
servers:
  - url: http://localhost:5000
...
paths:
...
  /assignments/{assignmentId}:
    get:
      tags:
        - Assignment
      summary: Haalt een specifieke opdracht op op basis van het assignment ID
      operationId: getAssignmentsById
      parameters:
        - name: assignmentId
          in: path
          required: true
          description: De unieke identifier van de opdracht
          schema:
            type: integer

      responses:
        '200':
          description: Details van de opdracht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          description: Ongeldige invoer, assignment ID is vereist
        '500':
          description: Serverfout bij het ophalen van de opdracht
...
components:
  schemas:
...
    LearningPath:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hruid:
          type: string
        title:
          type: string
        language:
          type: string
        description:
          type: string
        image:
          type: string
          nullable: true
          description: "Base64-gecodeerde cover afbeelding"
        num_nodes:
          type: integer
          nullable: true
          description: "Aantal nodes in het leerpad"
        num_nodes_left:
          type: integer
          nullable: true
          description: "Aantal resterende nodes voor de leerling"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        creatorId:
          type: integer
        creator:
          type: object
          properties:
            userId:
              type: integer
        nodes:
          type: array
          items:
            type: object
        assignments:
          type: array
          items:
            type: object
        questionGeneral:
          type: array
          items:
            type: object
...
    Question:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        title:
          type: string
        assignmentId:
          type: integer
        type:
          $ref: '#/components/schemas/QuestionType'
        createdAt:
          type: string
        updatedAt:
          type: string
    QuestionType:
      type: string
      enum: [SPECIFIC, GENERAL]
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        learningPathId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        deadline:
          type: string
        learningPath:
          $ref: '#/components/schemas/LearningPath'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        classAssignments:
          type: array
          items:
            $ref: '#/components/schemas/ClassAssignment'
        teamAssignments:
          type: array
          items:
            $ref: '#/components/schemas/TeamAssignment'
        submissions:
          type: array
    ClassAssignment:
      type: object
      properties:
        classId:
          type: integer
        assignmentId:
          type: integer
    TeamAssignment:
      type: object
      properties:
        teamId:
          type: integer
        assignmentId:
          type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT